requirements.txt
streamlit
opencv-python-headless
numpy
matplotlib
pillow
